refactor(api): init and generate code per module
