+++ api/types/tools.go
@@ -1,0 +1,8 @@
+//go:build tools
+// +build tools
+
+package tools
+
+import (
+	_ "github.com/deepmap/oapi-codegen/v2/cmd/oapi-codegen"
+)
